streets_service_base: change RapidJSON dependency
The RapidJSON library was previously built from source as a subproject
using FetchContent. However, while testing installation to non-standard
locations, it was discovered that overriding RapidJSON's install
location is not easy (if you want to do it properly). RapidJSON build
architecture does not seem conducive to subprojects, so it is probably
better to let the developer handle its building/installation.

Note: We should add a notice to builders that they need to install
RapidJSON from source using a version beyond a specific commit. This is
because the streets_service_base uses functionality not found in the
latest release.