Add rapidjson commit hash to clone commands
CARMA Streets uses functionality that was added after the 1.1.0 release,
but the project has not released a newer version. Additionally, building
the project from source will give it a 1.1.0 version even though it's
not actually that version. This can be confusing for CARMA Streets
consumers. As a (patchy) solution, a recent commit hash has been
specified for all RapidJSON git clone commands.